#6462 - BE - Integrate application with the DB schema provided
- Connection to dev SQL Server database
- Mapping entities, renaming columns
- Removing password, title and language from user
- Modifying JPA relations of roles and permissions to match with the new DB
- Disable token authentication temporally

+++ src/main/java/com/mcmcg/ico/bluefin/security/model/SecurityUser.java
@@ -16,7 +16,7 @@ import lombok.Data;
 public class SecurityUser implements UserDetails {
 
     private static final long serialVersionUID = 1680188661109204234L;
-    private Integer id;
+    private long id;
     private String username;
     private String password;
     private String email;
@@ -31,7 +31,7 @@ public class SecurityUser implements UserDetails {
         super();
     }
 
-    public SecurityUser(Integer id, String username, String password, String email,
+    public SecurityUser(long id, String username, String password, String email,
             Collection<? extends GrantedAuthority> authorities) {
         this.setId(id);
         this.setUsername(username);
